% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_code.R
\name{clean_data}
\alias{clean_data}
\title{Clean raw data collected with load_data function}
\usage{
clean_data(df, storm = "KATRINA-2005", dt = ymd_h("2005-08-29 12"),
  westernHemisphere = TRUE)
}
\arguments{
\item{df}{The raw data frame collected with load_data}

\item{storm}{The storm_id that we are filtering for}

\item{dt}{A datetime formatted in lubridate's ymd_h}

\item{westernHemisphere}{Format longitude to ensure it has negative values for locations in the Western hemisphere}
}
\value{
Returns a data frame
}
\description{
This will take our raw data, filter it for a particular storm observation and
modify it to the requested assignment format.
}
\examples{
\dontrun{
  df_raw <- load_data()
  katrina_observation <- clean_data(df_raw)
  print(katrina_observation)

  ike_observation <- clean_data(df_raw, storm="IKE-2008", dt=ymd_h("2008-09-13 6"))
  print(ike_observation)
}

}
