% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_code.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{The geom_hurricane function}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Used in conjunction with GeomHurricane.
See the layout of course materials p.g. 452
}
\examples{
\dontrun{
  geom_hurricane(data = storm_observation,
  aes(x = longitude, y = latitude, r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
  fill = wind_speed, colour = wind_speed), scale_radii = scale_radii)
}

}
