% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_code.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Defines Our GeomHurricane class}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\description{
As per the course materials:
The ggproto() function is used to construct a new class corresponding to your new geom.
The geom_* function is constructed as a regular function.
As a hint, notice that the wind radii geom essentially shows a polygon for each of the wind levels.
So we are interested in the polygonGrob.
As per coursera, we use the package
\href{https://cran.r-project.org/web/packages/geosphere/vignettes/geosphere.pdf}{geosphere}
}
\examples{
\dontrun{
  geom_hurricane(data = storm_observation,
  aes(x = longitude, y = latitude, r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
  fill = wind_speed, colour = wind_speed), scale_radii = scale_radii)
}

}
\keyword{datasets}
