% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_code.R
\name{load_data}
\alias{load_data}
\title{Load raw hurricane data from ebtrk_atlc_1988_2015.txt file}
\usage{
load_data()
}
\value{
Returns a data frame
}
\description{
This function was provided by Coursera to load data from the provided file ebtrk_atlc_1988_2015.txt
}
\examples{
\dontrun{
  df_raw <- load_data()
  head(df)
}

}
