% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_code.R
\name{map_observation}
\alias{map_observation}
\title{Map a particular storm observation}
\usage{
map_observation(storm_observation, location = "Louisiana", zoom = 6,
  scale_radii = 1, save_png = FALSE)
}
\arguments{
\item{storm_observation}{A dataframe collected with clean_data representing a storm observation}

\item{location}{A character string passed to get_map for the location we are looking to map}

\item{zoom}{The zoom used in ggmap::get_map. Default is 6.}

\item{scale_radii}{Numeric value which allows us to plot wind radii charts with radii scaled back.}

\item{save_png}{Boolean value specifying whether to save the resulting map into your working directory. Default is FALSE.}
}
\description{
This will take a storm observation gathered with the clean_data function
and map this observation.
}
\examples{
\dontrun{
df_raw <- load_data()
katrina_observation <- clean_data(df_raw) # get the default storm data
map_observation(katrina_observation, scale_radii = 1, save_png=TRUE)
}

}
